requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2024.12.14
Django==3.2.12
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
exceptiongroup==1.2.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.17.3
jsonschema-specifications
packaging
pluggy
pytest
pytest-django
pytz
PyYAML
referencing
sentry-sdk
sqlparse
tomli
typing_extensions
uritemplate
urllib3
pre-commit
Faker
factory_boy